Restructured motif code and tests
