+++ stumpy/gpu_ostinato.py
@@ -77,10 +77,8 @@ def gpu_ostinato(Ts, m, device_id=0, normalize=True):
         Ts, m, M_Ts, Σ_Ts, device_id=device_id, mp_func=gpu_stump
     )
 
-    (
-        central_radius,
-        central_Ts_idx,
-        central_subseq_idx,
-    ) = _get_central_motif(Ts, bsf_radius, bsf_Ts_idx, bsf_subseq_idx, m, M_Ts, Σ_Ts)
+    (central_radius, central_Ts_idx, central_subseq_idx,) = _get_central_motif(
+        Ts, bsf_radius, bsf_Ts_idx, bsf_subseq_idx, m, M_Ts, Σ_Ts
+    )
 
     return central_radius, central_Ts_idx, central_subseq_idx
