+++ stumpy/scrump.py
@@ -111,12 +111,7 @@ def _prescrump(
                 + T_B[i + k + m - 1] * T_A[j + k + m - 1]
             )
             D_squared = core._calculate_squared_distance(
-                m,
-                QT_j,
-                M_T[i + k],
-                Σ_T[i + k],
-                μ_Q[j + k],
-                σ_Q[j + k],
+                m, QT_j, M_T[i + k], Σ_T[i + k], μ_Q[j + k], σ_Q[j + k],
             )
             if D_squared < P_squared[i + k]:
                 P_squared[i + k] = D_squared
@@ -128,12 +123,7 @@ def _prescrump(
         for k in range(1, min(s, i + 1, j + 1)):
             QT_j = QT_j - T_B[i - k + m] * T_A[j - k + m] + T_B[i - k] * T_A[j - k]
             D_squared = core._calculate_squared_distance(
-                m,
-                QT_j,
-                M_T[i - k],
-                Σ_T[i - k],
-                μ_Q[j - k],
-                σ_Q[j - k],
+                m, QT_j, M_T[i - k], Σ_T[i - k], μ_Q[j - k], σ_Q[j - k],
             )
             if D_squared < P_squared[i - k]:
                 P_squared[i - k] = D_squared
